CAMEL-19398: cleanup type statistics
Move the statistics code to separate classes and interfaces
